requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
feedparser==6.0.11
python-dotenv==1.0.1
python-dateutil==2.9.0
praw==7.7.1
lxml==5.2.1
supabase>=2.3.0
